+++ main.tf
@@ -6,6 +6,8 @@ module "kms" {
   key_alias  = var.kms_key_alias == null ? "${var.namespace}-kms-alias" : var.kms_key_alias
   key_policy = var.kms_key_policy
 
+  policy_administrator_arn = var.kms_key_policy_administrator_arn
+
   create_clickhouse_key = var.enable_clickhouse
   clickhouse_key_alias  = var.kms_clickhouse_key_alias == null ? "${var.namespace}-kms-clickhouse-alias" : var.kms_clickhouse_key_alias
   clickhouse_key_policy = var.kms_clickhouse_key_policy
@@ -265,8 +267,9 @@ module "wandb" {
     module.app_eks,
     module.redis,
   ]
-  controller_image_tag   = "1.12.0"
-  operator_chart_version = "1.2.4"
+
+  operator_chart_version = var.operator_chart_version
+  controller_image_tag   = var.controller_image_tag
 
   spec = {
     values = {